requirements.txt
# requirements.txt
Django>=5
django-ninja>=1
psycopg2-binary>=2.9.9
redis>=5.0.1
celery>=5.3.4
django-cors-headers>=4.3.1
Pillow
python-decouple
django-phonenumber-field
phonenumbers
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
PyJWT>=2.8.0
dj-database-url>=2.1.0


# Development and testing
pytest~=7.4.3
pytest-django~=4.7.0
pytest-cov~=4.1.0
factory-boy~=3.3.0

# Code quality
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0

# Type hints for Django
django-stubs==4.2.7
types-redis==4.6.0.11

django-redis>=5.4.0
sentry-sdk>=1.44.0
argon2-cffi>=23.1.0
django-simple-history>=3.4.0
django-sequences>=2.7
twilio>=9.0.0
django-anymail>=10.3
django-ratelimit==4.1.0
reportlab>=4.0.9
django-qrcode
sentry_sdk
